fix: fix all test errors

+++ src/modules/artist/artist.service.ts
@@ -1,4 +1,4 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 // import { InMemoryUserStore } from '../../store/in-memory-user-store';
@@ -48,9 +48,16 @@ export class ArtistService {
   }
 
   async remove(id: string) {
-    validateId404(id, await this.prisma.artist.findMany());
-    return await this.prisma.artist.delete({ where: { id } });
+    // validateId404(id, await this.prisma.artist.findMany());
+    // return await this.prisma.artist.delete({ where: { id } });
 
+    const isArtistFavs = await this.prisma.favorites.findFirst({ where: { id } });
+    if (isArtistFavs) await this.prisma.favorites.delete({ where: { id } });
+
+    const isArtist = await this.prisma.artist.findFirst({ where: { id } });
+    if (isArtist) return await this.prisma.artist.delete({ where: { id } });
+
+    throw new NotFoundException();
     // this.store.tracks.forEach((track, index) => {
     //   if (track.artistId === id) this.store.tracks[index].artistId = null;
     // });
